api/requirements.txt
anyio==3.6.2
astroid==2.15.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
click==8.1.3
dill==0.3.6
exceptiongroup==1.1.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pika==1.3.1
platformdirs==3.1.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.5
pylint==2.16.4
pytest==7.2.2
pytest-dotenv==0.5.2
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
